capi: allow loading layouts and hotkey configs from files.
This reduces the amount of boilerplate code needed to do this from C.
The functions to deal with the raw file descriptors were extracted out
of `run.rs` and moved into `lib.rs`.
